Testing: introduce hound for acceptance testing
This uses Selenium with Firefox by default. We need to download selenium
and have it running in order to run the tests. Download `geckodriver`
from [here][1] and put it somewhere in your `PATH`. Then download the
"Selenium Standalone Server" from [here][2] and run it like so:

```sh
java -jar selenium-server-standalone-3.141.59.jar
```

This process isn't perfect, so hopefully we can improve it over time. It
is currently known to work with `geckodriver` 0.24.0 and
`selenium-server-standalone` 3.141.59.

[1]: https://github.com/mozilla/geckodriver/releases
[2]: https://www.seleniumhq.org/download/